ecommerce_project/requirement.txt
asgiref==3.7.2
Django==5.0
django-ckeditor==6.7.0
django-js-asset==2.2.0
mysqlclient==2.2.1
sqlparse==0.4.4
tzdata==2023.3
